requirements.txt
streamlit
google-cloud-language
beautifulsoup4
requests
pandas
